pyproject.toml
[tool.poetry]
name = "telegram-message-filter-bot"
version = "0.1.0"
description = "async telegram message filter bot"
authors = ["Mikhail Kolesov <m.m.kolesov@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
poetry = "^1.5.1"
python-telegram-bot = "^20.3"
pydantic = "^1.10.9"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
flake8 = "^6.0.0"
isort = "^5.12.0"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
default_section = "THIRDPARTY"
sections = ["STDLIB", "FIRSTPARTY", "THIRDPARTY", "LOCALFOLDER"]
known_local_folder = ["src", "bot", "core"]
lines_after_imports = "2"
skip = ["tests", ".venv", "venv"]
line_length = "79"
multi_line_output = "3"
include_trailing_comma = "true"
use_parentheses = "true"

[tool.flake8]
max-line-length = "79"
paths = ["./bot", "./core", "./src"]
max-complexity = "10"
exclude = ["tests", ".venv", "venv"]
per-file-ignores = [
    "__init__.py:F401",
]
